requirements.txt
alembic==1.5.8
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
greenlet==1.0.0
idna==2.10
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.12
sqlparse==0.4.1
urllib3==1.26.4
